third_party/WebKit/LayoutTests/fast/dom/shadow/focus-navigation-skips-non-focusable-shadow-in-iframe-expected.txt

CONSOLE WARNING: line 6: Element.createShadowRoot is deprecated and will be removed in M73, around March 2019. Please use Element.attachShadow instead. See https://www.chromestatus.com/features/4507242028072960 for more details.
Tests whether focus can navigate between the two buttons by pressing tab twice.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

PASS document.activeElement is iframe
PASS iframeDocument.activeElement is input
PASS document.activeElement is button2
PASS document.activeElement is iframe
PASS iframeDocument.activeElement is input
PASS document.activeElement is button1
PASS successfullyParsed is true

TEST COMPLETE

  
